fix: select usage
