+++ packages/settings-panel/src/presets/modals/add-members/index.ts
@@ -1,0 +1,1 @@
+export * from "./add-members";
